Replace logmain with logwriter
Implement initstatic_logwriter which gets a pointer to static memory
as its parameter.

Add a static version of logwriter to thread_localstore_t which is
used during init_threadcontext/free_threadcontext.

This replaces module logmain which was used during init/free.

Add TODO as preparation for new coming initlog module which is used
even before the first thread_localstore_t was initialized for the
main thread.

Fixed off-by-one error in allocatedsize_malloc.

Prepared valuecache / sys_pagesize_vm() for coming osinfo struct.

+++ C-kern/api/memory/vm.h
@@ -61,7 +61,7 @@ ramsize_t sizeavailableram_vm(void);
  * Returns the virtual memory page size supported by the underlying system.
  * This functions always calls the underlying system function.
  * The returned value is a power of two. */
-uint32_t sys_pagesize_vm(void);
+size_t sys_pagesize_vm(void);
 
 /* function: ismapped_vm
  * Returns true in case vmpage is mapped with accessmode equaling protection. */
